requirements.txt
# core
pydantic>=2.0
httpx>=0.27
python-dotenv>=1.0
tenacity>=8.2
nest_asyncio>=1.5
pyyaml>=6.0  # fleet/runtime/sampling YAML loaders

# data & compute
numpy>=1.26
scipy>=1.12
pandas>=2.2
pyarrow>=14  # LogProbTrace Parquet persistence (src/r5a/operators/p_logprob.py)
safetensors>=0.4  # WS6 hidden-state extraction (src/r5a/backends/offline_hf.py)

# visualization
matplotlib>=3.8
seaborn>=0.13

# notebook
jupyter>=1.0
ipykernel>=6.29
nbconvert>=7.0
tqdm>=4.66

# prompt optimization (notebook 06)
dspy>=2.5
